Updated average rating to use an external function, thus making code more modular
